+++ crates/maple_cli/src/commands/cache.rs
