+++ crates/maple_cli/src/commands/dumb_jump.rs
