+++ crates/maple_cli/src/commands/exec.rs
