+++ crates/maple_cli/src/commands/filter.rs
