+++ crates/maple_cli/src/commands/grep.rs
@@ -14,7 +14,7 @@ use icon::IconPainter;
 use utility::is_git_repo;
 
 use crate::app::Params;
-use crate::cmd::cache::{cache_exists, send_response_from_cache, SendResponse};
+use crate::commands::cache::{cache_exists, send_response_from_cache, SendResponse};
 use crate::process::light::{set_current_dir, LightCommand};
 use crate::tools::ripgrep::Match;
