+++ crates/maple_cli/src/lib.rs
@@ -5,7 +5,7 @@ mod process;
 mod stdio_server;
 mod tools;
 
-pub mod cmd;
+pub mod commands;
 /// Re-exports.
 pub use {
     anyhow::{Context, Result},
